requirements.txt
# 核心通信与 AI
openai>=1.0.0
roslibpy>=1.3
python-dotenv>=1.0
edge-tts>=6.0
requests>=2.30.0

# 硬件控制 (树莓派)
adafruit-circuitpython-neopixel
RPi.GPIO

# 基础数学与多媒体
numpy
opencv-python-headless
